added actionlist_template to display action list consistently on different pages added upload() method to enable myops to collect bootmanager and other logs added pcu error messages to policy setup clarified messages in emailTxt.py added log_path field to ActionRecord model to save relative bm log paths added SQL upgrade script for new actionrecord field distinguish between bm.log collection form bootman.py and directly uploaded logs removed unnecessary bootmanager action logs when nothing is done. fixed a bug in determining whether comon's dir was running on a node

+++ monitor/database/info/interface.py
@@ -114,7 +114,7 @@ class SiteInterface(HistorySiteRecord):
 		#       those errors.
 		
 		args = {'loginbase' : self.db.loginbase, 
-				'penalty_level' : self.db.penalty_level,
+				'penalty_level' : -self.db.penalty_level,
 				'monitor_hostname' : config.MONITOR_HOSTNAME,
 				'support_email'   : config.support_email,
 				'plc_name' : config.PLC_NAME,
@@ -201,7 +201,8 @@ class SiteInterface(HistorySiteRecord):
 							action_type='bootmanager_restore',
 							error_string="")
 
-		act = ActionRecord(loginbase=self.db.loginbase,
+		if ret:
+			act = ActionRecord(loginbase=self.db.loginbase,
 							hostname=hostname,
 							action='reboot',
 							action_type='bootmanager_' + ret,
